New Default Widget!
The default widget is the widget, which will appear when there there a no Widgets in the Queue. Mean, will act as an "ScreenSaver" / "Fallback" Widget! It must be define.
if a External Managet (like the Menu Widget by pressing a Menu button) or StatusWidgets (i.e. Programming Mode button) are are managed with DisplayEnabled, they will be shown at first and also the Normal Widgets, if the Queue is not Empty! The Timer can be configured. The current inactivity default time is, 1 Sec.